+++ students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/Task3.java
@@ -1,4 +1,4 @@
-package ru.mirea.practice.s21K0689;
+package ru.mirea.practice;
 
 abstract class Task3 {
     public static void main(String[] args) {
