+++ students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/Task4.java
@@ -1,4 +1,4 @@
-package ru.mirea.practice.s21K0689;
+package ru.mirea.practice;
 
 abstract class Task4 {
     public static void main(String[] args) {
