+++ students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/Task5.java
@@ -1,10 +1,12 @@
-package ru.mirea.practice.s21K0689;
+package ru.mirea.practice;
 
 abstract class Task5 {
     public static int fact(int x) {
         if (x > 1) {
             return x * fact(x - 1);
-        } else return 1;
+        } else {
+            return 1;
+        }
     }
 
     public static void main(String[] args) {
