--- students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/s21K0689/Task1.java
@@ -1,16 +1,0 @@
-package ru.mirea.practice.s21K0689;
-
-abstract class Task1 {
-    public static void main(String[] args) {
-        int[] a = {1, 23, 4, 235, 246, 123};
-        double sum = 0;
-        double average;
-        for (int j : a) {
-            System.out.println(j);
-            sum += j;
-        }
-        average = sum / a.length;
-        System.out.println("sum = " + sum);
-        System.out.println("average = " + average);
-    }
-}
